pyproject.toml
[project]
name = "mcp-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.11.2",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    # Backend API stack
    "fastapi>=0.110.0",
    "uvicorn>=0.30.0",
    "SQLAlchemy>=2.0.30",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "redis>=5.0.0",
    # HTTP client used across services
    "httpx>=0.27.0",
]
